Simple Dice Game Simulation
R Code that simulates the probability of a two players who play three rounds of a dice game.  The first player rolls the his die and records the outcome.  The second player does the same.  If the dice match between the players the Round is "Won."  If all three rounds are "won" the entire game is won.  The odds of winning are then computed.  Note in this particular simulation, player two always follows and attempts to match player one's result.